+++ test/test_generic_commands.py
@@ -3,7 +3,7 @@ from time import sleep, time
 
 import pytest
 import redis
-
+from redis.exceptions import ResponseError
 import testtools
 
 
@@ -321,6 +321,20 @@ def test_expire_should_expire_key(r):
 
 
 @testtools.run_test_if_redispy_ver('above', '4.2.0')
+def test_expire_should_throw_error(r):
+    r.set('foo', 'bar')
+    assert r.get('foo') == b'bar'
+    with pytest.raises(ResponseError):
+        r.expire('foo', 1, nx=True, xx=True)
+    with pytest.raises(ResponseError):
+        r.expire('foo', 1, nx=True, gt=True)
+    with pytest.raises(ResponseError):
+        r.expire('foo', 1, nx=True, lt=True)
+    with pytest.raises(ResponseError):
+        r.expire('foo', 1, gt=True, lt=True)
+
+
+@testtools.run_test_if_redispy_ver('above', '4.2.0')
 @pytest.mark.max_server('7')
 def test_expire_extra_params_return_error(r):
     with pytest.raises(redis.exceptions.ResponseError):
@@ -486,3 +500,39 @@ def test_set_float_value(r):
     x = 1.23456789123456789
     r.set('foo', x)
     assert float(r.get('foo')) == x
+
+
+@pytest.mark.min_server('7')
+@testtools.run_test_if_redispy_ver('above', '4.2.0')
+def test_expire_should_not_expire__when_no_expire_is_set(r):
+    r.set('foo', 'bar')
+    assert r.get('foo') == b'bar'
+    assert r.expire('foo', 1, xx=True) == 0
+
+
+@pytest.mark.min_server('7')
+@testtools.run_test_if_redispy_ver('above', '4.2.0')
+def test_expire_should_not_expire__when_expire_is_set(r):
+    r.set('foo', 'bar')
+    assert r.get('foo') == b'bar'
+    assert r.expire('foo', 1, nx=True) == 1
+    assert r.expire('foo', 2, nx=True) == 0
+
+
+@pytest.mark.min_server('7')
+@testtools.run_test_if_redispy_ver('above', '4.2.0')
+def test_expire_should_expire__when_expire_is_greater(r):
+    r.set('foo', 'bar')
+    assert r.get('foo') == b'bar'
+    assert r.expire('foo', 100) == 1
+    assert r.get('foo') == b'bar'
+    assert r.expire('foo', 200, gt=True) == 1
+
+
+@pytest.mark.min_server('7')
+@testtools.run_test_if_redispy_ver('above', '4.2.0')
+def test_expire_should_expire__when_expire_is_lessthan(r):
+    r.set('foo', 'bar')
+    assert r.get('foo') == b'bar'
+    assert r.expire('foo', 20) == 1
+    assert r.expire('foo', 10, lt=True) == 1
